pyproject.toml
[tool.poetry]
name = "Social Network"
version = "0.0.1"
description = "Social Network FastAPI"
authors = ["EscapeFromHell"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.6"
fastapi = ">=0.85.0"
uvicorn = ">=0.18.3"
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
psycopg2 = "^2.9.3"
pydantic = {extras = ["email"], version = "^1.10.2"}
httpx = ">=0.23.0"
requests = "^2.28.1"
python-multipart = "~0.0.5"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
urllib3 = "1.26.16"
python-dotenv = "^0.21.0"

black = ">=22.1.0"
isort = ">=5.9.3"

flake8 = ">=3.9.2"
flake8-broken-line = ">=0.3.0"
flake8-bugbear = ">=21.9.1"
flake8-builtins = ">=1.5.3"
flake8-coding = ">=1.3.2"
flake8-comprehensions = ">=3.6.1"
flake8-debugger = ">=4.0.0"
flake8-eradicate = ">=1.1.0"
flake8-logging-format = ">=0.6.0"
flake8-pep3101 = ">=1.3.0"
flake8-print = ">=4.0.0"
flake8-pytest-style = ">=1.5.0"
flake8-string-format = ">=0.3.0"
flake8-variables-names = ">=0.0.4"
flake8-expression-complexity = ">=0.0.9"
flake8-cognitive-complexity = ">=0.1.0"
flake8-annotations = ">=2.6.2"
flake8-noqa = ">=1.1.0"
flake8-docstrings = ">=1.6.0"
flake8-unused-arguments = ">=0.0.6"
pep8-naming = ">=0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = '''
(
  /(
      \.git         # exclude a few common directories in the
    | \.pytest_cache          # root of the project
    | \.idea
    | \__pycache__
    | \.venv
    | alembic
    | env
    | venv
  )/
)
'''

[tool.isort]
profile = "black"
line_length = 110
multi_line_output = 3
include_trailing_comma = true
skip = [".git", ".pytest_cache", ".idea", "__pycache__", ".venv", "alembic", "env", "venv"]
